+++ src/schemas/contribution.ts
@@ -1,4 +1,4 @@
-import { FileType, ReleaseType } from '@prisma/client';
+import { Bitrate, FileType, ReleaseMedia, ReleaseType } from '@prisma/client';
 import { z } from 'zod';
 
 const releaseTypeEnum = z.enum(
@@ -6,6 +6,10 @@ const releaseTypeEnum = z.enum(
 );
 const fileTypeEnum = z.enum(
   Object.values(FileType) as [FileType, ...FileType[]]
+);
+const bitrateEnum = z.enum(Object.values(Bitrate) as [Bitrate, ...Bitrate[]]);
+const mediaEnum = z.enum(
+  Object.values(ReleaseMedia) as [ReleaseMedia, ...ReleaseMedia[]]
 );
 
 export const createContributionSchema = z.object({
@@ -34,16 +38,8 @@ export const createContributionSchema = z.object({
     .string()
     .optional()
     .transform((value) => value?.trim() || undefined),
-  bitrate: z
-    .string()
-    .max(50)
-    .optional()
-    .transform((value) => value?.trim() || undefined),
-  media: z
-    .string()
-    .max(50)
-    .optional()
-    .transform((value) => value?.trim() || undefined),
+  bitrate: bitrateEnum.optional(),
+  media: mediaEnum.optional(),
   hasLog: z.boolean().optional().default(false),
   hasCue: z.boolean().optional().default(false),
   isScene: z.boolean().optional().default(false),
@@ -67,16 +63,8 @@ export const addContributionToReleaseSchema = z.object({
     .string()
     .optional()
     .transform((value) => value?.trim() || undefined),
-  bitrate: z
-    .string()
-    .max(50)
-    .optional()
-    .transform((value) => value?.trim() || undefined),
-  media: z
-    .string()
-    .max(50)
-    .optional()
-    .transform((value) => value?.trim() || undefined),
+  bitrate: bitrateEnum.optional(),
+  media: mediaEnum.optional(),
   hasLog: z.boolean().optional().default(false),
   hasCue: z.boolean().optional().default(false),
   isScene: z.boolean().optional().default(false)